Move column classes from core-db to core-api for reuse.
