+++ transmart-core-api/src/main/groovy/org/transmartproject/core/dataquery/ColumnDataType.groovy
@@ -1,0 +1,8 @@
+package org.transmartproject.core.dataquery
+
+/**
+ * A data type of column values.
+ */
+enum ColumnDataType {
+    NUMERIC, DATE, STRING
+}
