+++ transmart-core-api/src/main/groovy/org/transmartproject/core/dataquery/Measure.groovy
@@ -1,0 +1,8 @@
+package org.transmartproject.core.dataquery
+
+/**
+ * Levels of measurement. Used in the SPSS export.
+ */
+enum Measure {
+    NOMINAL, ORDINAL, SCALE
+}
