+++ transmart-core-api/src/main/groovy/org/transmartproject/core/dataquery/MetadataAwareDataColumn.groovy
@@ -1,0 +1,8 @@
+package org.transmartproject.core.dataquery
+
+/**
+ * A data column that contains metadata about itself.
+ */
+interface MetadataAwareDataColumn extends DataColumn {
+    ColumnMetadata getMetadata()
+}
