+++ transmart-core-db/src/integration-test/groovy/org/transmartproject/db/multidimquery/SubjectObservationsByStudyConceptsTableViewSpec.groovy
@@ -9,11 +9,11 @@ import org.transmartproject.db.multidimquery.query.Constraint
 import org.transmartproject.db.multidimquery.query.StudyNameConstraint
 import org.transmartproject.db.user.User
 import spock.lang.Specification
-import static org.transmartproject.db.multidimquery.ColumnDataType.NUMERIC
-import static org.transmartproject.db.multidimquery.ColumnDataType.STRING
-import static org.transmartproject.db.multidimquery.ColumnDataType.DATE
-import static org.transmartproject.db.multidimquery.Measure.NOMINAL
-import static org.transmartproject.db.multidimquery.Measure.SCALE
+import static org.transmartproject.core.dataquery.ColumnDataType.NUMERIC
+import static org.transmartproject.core.dataquery.ColumnDataType.STRING
+import static org.transmartproject.core.dataquery.ColumnDataType.DATE
+import static org.transmartproject.core.dataquery.Measure.NOMINAL
+import static org.transmartproject.core.dataquery.Measure.SCALE
 
 @Rollback
 @Integration
